Revert "feat: [#448] watch iphoone WorkoutManager 통합"
