--- SoccerBeat/Common/Utiles/View+extension.swift
@@ -1,34 +1,0 @@
-//
-//  View+extension.swift
-//  SoccerBeat
-//
-//  Created by jose Yun on 10/25/24.
-//
-
-import Foundation
-import SwiftUI
-
-extension View where Self: Shape {
-  func glow(
-    fill: some ShapeStyle,
-    lineWidth: Double,
-    blurRadius: Double = 8.0,
-    lineCap: CGLineCap = .round
-  ) -> some View {
-    self
-      .stroke(style: StrokeStyle(lineWidth: lineWidth / 2, lineCap: lineCap))
-      .fill(fill)
-      .overlay {
-        self
-          .stroke(style: StrokeStyle(lineWidth: lineWidth, lineCap: lineCap))
-          .fill(fill)
-          .blur(radius: blurRadius)
-      }
-      .overlay {
-        self
-          .stroke(style: StrokeStyle(lineWidth: lineWidth, lineCap: lineCap))
-          .fill(fill)
-          .blur(radius: blurRadius / 2)
-      }
-  }
-}
