+++ SoccerBeat/SoccerBeat Watch App/Sources/Features/EndRecording/SummaryView.swift
@@ -70,6 +70,7 @@ struct SummaryView: View {
                         .onAppear {
                             withAnimation {
                                 isHeartBlinking = true
+                                print("heart", isHeartBlinking)
                             }
                          }
                     }
