+++ SoccerBeat/SoccerBeat Watch App/Sources/WorkoutManager/WorkoutManager+extensions.swift
@@ -12,122 +12,6 @@ import OSLog
 
 // MARK: - 세션 pause, resume, end 시에 작동하는 델리게이트
 extension WorkoutManager: HKWorkoutSessionDelegate {
-    
-    // MARK: - 데이터 수집 및 경기 시작
-    func startWorkout() {
-        setupWorkoutConfig()
-        startWorkoutSession()
-    }
-    
-    // 워치 경기 기록 설정
-    private func setupWorkoutConfig() {
-        // workout configuration 설정
-        let configuration = HKWorkoutConfiguration()
-        configuration.activityType = .running
-        configuration.locationType = .outdoor
-
-        // 세션, 빌더, 루트 빌더, 로케이션 매니저 초기화
-        do {
-            session = try HKWorkoutSession(
-                healthStore: healthStore,
-                configuration: configuration
-            )
-        } catch {
-            NSLog(error.localizedDescription)
-        }
-        builder = session?.associatedWorkoutBuilder()
-        routeBuilder = HKWorkoutRouteBuilder(healthStore: healthStore, device: .local())
-        // 델리게이트 선언
-        session?.delegate = self
-        builder?.delegate = self
-        builder?.dataSource = HKLiveWorkoutDataSource(healthStore: healthStore,
-                                                      workoutConfiguration: configuration)
-
-    }
-    
-    // 워치 경기 기록 시작
-    private func startWorkoutSession() {
-        
-        let startDate = Date()
-        session?.startActivity(with: startDate)
-        builder?.beginCollection(withStart: startDate) { (_, _) in
-            // The workout has started.
-        }
-        // 위치 정보 수집
-        locationManager.startUpdatingLocation()
-
-        // 헬스킷에서 나이 정보를 통해 적절한 최대심박수 찾기
-        matrics.computeProperMaxHeartRate(with: healthStore)
-        
-        // 워치 세션을 원격 세션과 연동
-        Task {
-            do {
-                try await session?.startMirroringToCompanionDevice()
-            } catch {
-                NSLog("Error in mirroring", error.localizedDescription)
-            }
-        }
-    }
-    
-    // 워치 경기 기록 종료
-    func endWorkoutSession(_ date: Date) async throws {
-        do {
-            try await builder?.endCollection(at: date)
-        } catch {
-            throw SessionError.failureBuilderEnd
-        }
-
-        guard let workout = try await builder?.finishWorkout() else {
-            throw SessionError.failureFinishWorkout
-        }
-        Task { @MainActor in
-            self.workout = workout
-        }
-
-        let metadata = self.matrics.getMetadata()
-
-        guard let route = try await routeBuilder?.finishRoute(
-            with: workout,
-            metadata: metadata
-        ) else {
-            throw SessionError.failureMakeRoute
-            
-        }
-        Task { @MainActor in
-            self.route = route
-        }
-    }
-    
-    // 워치 경기 기록 초기화
-    func resetWorkout() {
-        builder = nil
-        workout = nil
-        session = nil
-        
-        matrics.reset()
-    }
-    
-    func togglePause() {
-        running ? pause() : resume()
-    }
-    
-    private func pause() {
-        session?.pause()
-    }
-    
-    private func resume() {
-        session?.resume()
-    }
-    
-    func endWorkout() {
-        session?.end()
-        showingSummaryView.toggle()
-    }
-}
-
-// MARK: - WorkoutData 수집시 동작하는 델리게이트
-extension WorkoutManager: HKLiveWorkoutBuilderDelegate {
-    
     func workoutSession(_ workoutSession: HKWorkoutSession, didChangeTo toState: HKWorkoutSessionState,
                         from fromState: HKWorkoutSessionState, date: Date) {
         NSLog("WorkOutSession 변화 감지: \(toState)")
@@ -153,7 +37,10 @@ extension WorkoutManager: HKLiveWorkoutBuilderDelegate {
     func workoutSession(_ workoutSession: HKWorkoutSession, didFailWithError error: Error) {
         
     }
+}
 
+// MARK: - WorkoutData 수집시 동작하는 델리게이트
+extension WorkoutManager: HKLiveWorkoutBuilderDelegate {
     func workoutBuilderDidCollectEvent(_ workoutBuilder: HKLiveWorkoutBuilder) {
         
     }
@@ -174,7 +61,7 @@ extension WorkoutManager: HKLiveWorkoutBuilderDelegate {
 }
 
 // MARK: - 위치 정보 수집시 동작하는 델리게이트
-extension WorkoutManager {
+extension WorkoutManager: CLLocationManagerDelegate {
     // MARK: - 위치 정보가 수집되면 불리는 메서드
     func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
         
@@ -184,12 +71,15 @@ extension WorkoutManager {
             location.horizontalAccuracy <= 20.0
         }
         
+        // 성공치를 외부에 데이터로 넘기기, 혹은 에러 뭐시기를 하는게 좋으려나?
         guard !filteredLocations.isEmpty else {
+            // 실패해도 저장을 하네?
             routeBuilder?.insertRouteData(locations, completion: { _, _ in
             })
             return
         }
         
+        // 성공해도 저장을 하네?
         // Add the filtered data to the route.
         routeBuilder?.insertRouteData(filteredLocations) { (success, error) in
             if !success {
@@ -205,6 +95,7 @@ extension WorkoutManager {
     }
 
     func checkLocationAuthorization() {
+        hasLocationAuthorization = false
         switch locationManager.authorizationStatus {
         case .notDetermined:
             NSLog("위치 권한 결정 안됨")
@@ -216,6 +107,7 @@ extension WorkoutManager {
         case .authorizedAlways, .authorizedWhenInUse:
             NSLog("위치 권한 항상 허용 혹은 사용 중 허용")
             locationManager.startUpdatingLocation()
+            hasLocationAuthorization = true
         @unknown default:
             NSLog(locationManager.authorizationStatus.rawValue.description)
         }
