+++ SoccerBeat/SoccerBeat/Sources/Features/MatchRecords/MatchRecapView.swift
@@ -9,13 +9,14 @@ import SwiftUI
 import StoreKit
 
 struct MatchRecapView: View {
-    @EnvironmentObject var workoutManager: WorkoutManager
+    @EnvironmentObject var healthInteractor: HealthInteractor
     @State private var userName = ""
     @Binding var workouts: [WorkoutData]
     @State private var requestReview = false
     @State private var hasDoneReviewBefore = false
     private let reviewRequestThreshold: TimeInterval = 4 * 30 * 24 * 60 * 60 // 4 months
-        
+    
+    
     private var lastName: String {
         guard let lastName = userName
             .split(separator: " ")
@@ -145,7 +146,7 @@ struct MatchRecapView: View {
     
     private func delete(_ offset: IndexSet) async {
         do {
-            try await workoutManager.delete(at: offset)
+            try await healthInteractor.delete(at: offset)
         } catch {
             NSLog("Deleting HKWorkout failed")
         }
@@ -301,9 +302,7 @@ extension MatchListItemView {
 }
 
 #Preview {
-    @StateObject var workoutManager = DIContianer.makeWorkoutManager()
-    return
     MatchRecapView(workouts: .constant(WorkoutData.exampleWorkouts))
-        .environmentObject(ProfileModel(workoutManager: workoutManager))
-        .environmentObject(workoutManager)
+        .environmentObject(ProfileModel(healthInteractor: HealthInteractor()))
+        .environmentObject(HealthInteractor())
 }
